Merge release to master

+++ src/backend/league_service/src/main/java/ies/gamesense/league_service/entities/LeagueClubId.java
@@ -4,30 +4,30 @@ import java.io.Serializable;
 import java.util.Objects;
 
 public class LeagueClubId implements Serializable {
-    private int league;
-    private int club_id;
+    private long league;
+    private long club_id;
 
     public LeagueClubId() {}
 
-    public LeagueClubId(int league, int club_id) {
+    public LeagueClubId(long league, long club_id) {
         this.league = league;
         this.club_id = club_id;
     }
 
     // Getters, setters, equals, and hashCode methods
-    public int getLeague() {
+    public long getLeague() {
         return league;
     }
 
-    public void setLeague(int league) {
+    public void setLeague(long league) {
         this.league = league;
     }
 
-    public int getClub_id() {
+    public long getClub_id() {
         return club_id;
     }
 
-    public void setClub_id(int club_id) {
+    public void setClub_id(long club_id) {
         this.club_id = club_id;
     }
 
@@ -44,4 +44,3 @@ public class LeagueClubId implements Serializable {
         return Objects.hash(league, club_id);
     }
 }
-
